+++ lib/telemetrydecksdk.dart
@@ -1,4 +1,7 @@
 library;
 
+export 'src/error_category.dart';
+export 'src/purchase_event.dart';
+export 'src/purchase_type.dart';
 export 'src/telemetry_manager_configuration.dart';
 export 'src/telemetrydecksdk.dart';